backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
